Only parse whole templates
Without this, all partials are first run through the Formatter, thus the
predefined input tags are replaced with the desired output tags.
Then once all the partials are compiled into the main calling template,
it would call the Formatter again, this time stripping out the output
formatting already performed on the partials.

Now it just compiles all the partials into the main template, and then
runs the Formatter once on the complete source.